requirements.txt
flask_sqlalchemy==3.0.5
flask==2.3.3
google_api_python_client==2.97.0
flask_debugtoolbar==0.13.1
xlwings==0.30.11
openpyxl==3.1.2
SQLAlchemy~=2.0.20